% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{cleanFeatureSets}
\alias{cleanFeatureSets}
\title{Clean feature sets}
\usage{
cleanFeatureSets(x)
}
\arguments{
\item{x}{An IndexedRelations object.}
}
\value{
An IndexedRelations object with unique and sorted feature sets.
}
\description{
Clean feature sets within a \linkS4class{IndexedRelations} object.
}
\details{
This function will sort and remove duplicates within each feature set.
This provides some structure that can be useful in some algorithms, 
e.g., ordering of partner indices reflects the ordering of the partner features.
}
\examples{
library(GenomicRanges)
promoters <- GRanges("chrA", IRanges(1:10*20, 1:10*20+10))
enhancers <- GRanges("chrA", IRanges(1:20*10, 1:20*10+10))
partner1 <- sample(length(promoters), 100, replace=TRUE)
partner2 <- sample(length(enhancers), 100, replace=TRUE)

rel <- IndexedRelations(
    list(promoter=partner1, enhancer=partner2),
    featureSets=list(promoters=promoters, enhancers=enhancers)
)
rel <- cleanFeatureSets(rel)

featureSets(rel)[[1]]
featureSets(rel)[[2]]

}
\seealso{
\code{\link{standardizeFeatureSets}}, to do the same thing with \code{clean=TRUE}.
}
\author{
Aaron Lun
}
