% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\docType{methods}
\name{IR conversions}
\alias{IR conversions}
\alias{makePairsFromIndexedRelations}
\title{IndexedRelation conversions}
\description{
Convert an IndexedRelations object to and from different Bioconductor classes.
}
\details{
A number of \code{coerce} methods are available to convert to/from an \linkS4class{IndexedRelations} object:
\itemize{
\item \code{as(from, "IndexedRelations")} for a \linkS4class{Pairs} \code{x} will return an IndexedRelations object where the first vector in \code{from} becomes the first partner in the output.
\item \code{as(from, "IndexedRelations")} for a \linkS4class{DataFrame} \code{x} will return an IndexedRelations object where the first column in \code{from} becomes the first partner in the output.
\item \code{as(from, "DataFrame")} for an IndexedRelations \code{x} will return a DataFrame object where the first column is the first partner in \code{From}.
}

In addition, \code{makePairsFromIndexedRelations(x)} will create a Pairs instance from an IndexedRelations \code{x}.
This requires that \code{x} only has two partners in each relationship, otherwise an error is thrown.
}
\examples{
p <- Pairs(1:10, Rle(1L, 10), score=rnorm(10), names=letters[1:10])

ir <- as(p, "IndexedRelations")
ir

back <- makePairsFromIndexedRelations(ir)
back

df <- as(ir, "DataFrame")
df

as(df, "IndexedRelations")


}
\author{
Aaron Lun
}
