% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangePartners.R
\name{rearrangePartners}
\alias{rearrangePartners}
\title{Rearrange partners}
\usage{
rearrangePartners(x, perm)
}
\arguments{
\item{x}{An IndexedRelations object.}

\item{perm}{An integer vector specifying the permutation of the partners.}
}
\value{
An IndexedRelations object with rearranged partners.
}
\description{
Rearrange partners in an \linkS4class{IndexedRelations} object.
}
\details{
This function facilitates the task of switching the order of partners across all relationships in \code{x}.
}
\examples{
genomic <- GRanges("chrA", IRanges(1:10*20, 1:10*20+10))
intervals <- IRanges(1:20*10, 1:20*10+10))

rel <- IndexedRelations(list(1:2, 1:2), list(genomic, intervals))

rearrangePartners(rel, c(2, 1))

# Or even drop a partner completely:
rearrangePartners(rel, 2)

}
\author{
Aaron Lun
}
